Started using Swagger for generating API docs. Completed orders creation.

+++ server/models/ingredient.go
@@ -1,9 +1,10 @@
 package models
 
 // Ingredient ...
+// swagger:model
 type Ingredient struct {
 	ID       int     `json:"id,omitempty"`
 	Name     string  `json:"name,omitempty"`
 	Category string  `json:"type,omitempty"`
-	Price    float32 `json:"price,omitempty"`
+	Price    float64 `json:"price,omitempty"`
 }
